+++ apps/manifest-demo/3009-webpack-provider/src/components/WebpackPng.tsx
@@ -1,0 +1,5 @@
+import png from '../../public/webpack.png';
+
+export default function WebpackPng() {
+  return <img className="remote1-webpack-png" src={png} alt="webpack png" />;
+}